pyproject.toml
[tool.poetry]
name = "omni-dl"
version = "0.1.0"
description = "Omni Downloader"
authors = ["Ming Yang <ymviv@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "omni_dl"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.101.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
odmantic = "^0.9.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.5.2"
requests = "^2.31.0"
fsspec = "^2023.6.0"
pydantic = ">=1.0.0,<2.0.0"
python-multipart = "^0.0.6"
celery = {extras = ["redis"], version = "^5.3.1"}
s3fs = "^2023.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
